requirements.txt
streamlit
langchain
langchain-openai
faiss-cpu
sentence-transformers
numpy
python-dotenv
